feat: move unrelated functions from matcher to recorder module - move tests regarding changes
